moved repeated time conversion code into utility file
